pyproject.toml
[build-system]
requires = ["setuptools >= 61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spotify_to_tidal"
version = "1.0.4"
requires-python = ">= 3.10"

dependencies = [
  "spotipy~=2.24.0",
  "tidalapi==0.8.8",
  "pyyaml~=6.0",
  "tqdm~=4.64",
  "sqlalchemy~=2.0",
  "pytest~=8.0",
  "pytest-mock~=3.8"
]

[tools.setuptools.packages."spotify_to_tidal"]
where = "src"
include = "spotify_to_tidal*"

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]

[project.scripts]
spotify_to_tidal = "spotify_to_tidal.__main__:main"
